+++ include/tao/pegtl/internal/rules.hpp
@@ -36,6 +36,7 @@
 #include "not_at.hpp"
 #include "one.hpp"
 #include "opt.hpp"
+#include "opt_must.hpp"
 #include "pad.hpp"
 #include "pad_opt.hpp"
 #include "plus.hpp"